requirements.txt
python-telegram-bot==20.7
groq==0.4.2
anthropic==0.39.0
mysql-connector-python==8.2.0
python-dotenv==1.0.0
requests==2.31.0
